search: Move CommitSearchResult to internal/search/result
